+++ lifters-log-rewrite/src/features/auth/api/changePassword.ts
@@ -1,19 +1,18 @@
 import checkStatus from "../../../utils/api/checkStatus";
 import { SERVER_URL } from "../../../utils/constants";
 
-type NewUser = {
-    username: string;
-    password: string;
+type NewPasswordInfo = {
+    newPassword: string;
     email: string;
 }
 
-async function createNewUser(newUser: NewUser) {
-    const response = await fetch(`${SERVER_URL}/profile/new`, {
+async function changePassword(newInfo: NewPasswordInfo) {
+    const response = await fetch(`${SERVER_URL}/auth/changepassword`, {
         method: "POST",
-        body: JSON.stringify(newUser),
+        body: JSON.stringify(newInfo),
         headers: {"Content-type": "application/json"}
     });
     return await checkStatus(response);
 }
 
-export default createNewUser;
+export default changePassword