+++ lifters-log-rewrite/src/features/history/api/getHistoryForDate.ts
@@ -2,10 +2,10 @@ import { QueryFunctionContext } from "@tanstack/react-query";
 import checkStatus from "../../../utils/api/checkStatus";
 import { SERVER_URL } from "../../../utils/constants";
 
-async function getHistoryForDate({ queryKey }: QueryFunctionContext<[string, string, string, string]>) {
-    const [, , date, movement] = queryKey
+async function getHistoryForDate({ queryKey }: QueryFunctionContext<[string, string, string, number]>) {
+    const [, , date, exerciseid] = queryKey
 
-    const response = await fetch(`${SERVER_URL}/history/${movement}/${date}`, {
+    const response = await fetch(`${SERVER_URL}/history/${exerciseid}/${date}`, {
         method: "GET",
         credentials: "include"
     });