+++ lifters-log-rewrite/src/features/insights/api/getExerciseInfo.ts
@@ -2,13 +2,12 @@ import { QueryFunctionContext } from "@tanstack/react-query";
 import { SERVER_URL } from "../../../utils/constants";
 import checkStatus from "../../../utils/api/checkStatus";
 
-
-async function getMovementInsights({ queryKey }: QueryFunctionContext<[string, number, string]>) {
-    const [, exerciseid, timeframe] = queryKey
+async function getExericseInfo({ queryKey }: QueryFunctionContext<[string, number]>) {
+    const [, exerciseid] = queryKey
 
     if (exerciseid < 1) return;
 
-    const response = await fetch(`${SERVER_URL}/stats/${exerciseid}/${timeframe}`, {
+    const response = await fetch(`${SERVER_URL}/insights/exerciseinfo/${exerciseid}`, {
         method: "GET",
         credentials: "include"
     });
@@ -16,4 +15,4 @@ async function getMovementInsights({ queryKey }: QueryFunctionContext<[string, n
     return await checkStatus(response)
 }
 
-export default getMovementInsights
+export default getExericseInfo;