+++ lifters-log-rewrite/src/features/profile/components/EditUserProfile.tsx
@@ -5,10 +5,16 @@ import UserData from "../../../types/UserData";
 import { PLACEHOLDERUSERDATA } from "../../../utils/constants";
 import ProfilePictureChanger from "./ProfilePictureChanger";
 import getUserPreferences from "../api/getUserPreferences";
-import { useEffect, useState } from "react";
+import { ChangeEvent, useEffect, useState } from "react";
 import PreferenceMenu from "./PreferenceMenu";
 import { isMobile } from "react-device-detect";
 import uploadProfilePicture from "../api/uploadProfilePicture";
+import PopupMenu from "../../../components/PopupMenu";
+import updateEmail from "../api/updateEmail";
+import { checkEmail, checkUsername } from "../../../utils/checkStrings";
+import FadePopup from "../../../components/FadePopup";
+import updateUsername from "../api/updateUsername";
+import requestPasswordChange from "../api/requestPasswordChange";
 
 type EditUserProfileProps = {
     pfpurl: string;
@@ -27,6 +33,13 @@ function EditUserProfile({ pfpurl, cancelFn }: EditUserProfileProps) {
     const [xpAnimation, setXpAnimation] = useState(true)
     const [liftRecords, setLiftRecords] = useState(true)
 
+    const [changingUsername, setChangingUsername] = useState(false)
+    const [newUsername, setNewUsername] = useState("")
+    const [changingEmail, setChangingEmail] = useState(false)
+    const [newEmail, setNewEmail] = useState("")
+    const [updateMessage, setUpdateMessage] = useState("")
+    const [requestedPassword, setRequesedPassword] = useState(false);
+
     const [profileImage, setProfileImage] = useState<File | null>(null)
 
     const pfpSize = isMobile ? 128 : 256;
@@ -34,11 +47,41 @@ function EditUserProfile({ pfpurl, cancelFn }: EditUserProfileProps) {
     const preferenceMutation = useMutation({
         mutationFn: setPreferences
     })
-
     const newPicMutation = useMutation({
         mutationFn: uploadProfilePicture,
         onSuccess: () => {
             queryClient.invalidateQueries({ queryKey: ["stats"] })
+        }
+    })
+    const changeEmailMutation = useMutation({
+        mutationFn: updateEmail,
+        onSuccess: () => {
+            authUser.email = newEmail;
+            setUpdateMessage("Please check your email (and spam) to verify your account!")
+            setChangingEmail(false)
+        },
+        onError: (error: Error) => {
+            setUpdateMessage(error.message)
+        }
+    })
+    const changeUsernameMutation = useMutation({
+        mutationFn: updateUsername,
+        onSuccess: () => {
+            authUser.username = newUsername
+            setUpdateMessage("Username changed! Some areas may take time to update.")
+            setChangingUsername(false)
+        },
+        onError: (error: Error) => {
+            setUpdateMessage(error.message)
+        }
+    })
+    const changePasswordMutation = useMutation({
+        mutationFn: requestPasswordChange,
+        onSuccess: () => {
+            setUpdateMessage("Password change requested! Please check your email to finish changing your password.")
+        },
+        onError: (error: Error) => {
+            setUpdateMessage(error.message)
         }
     })
 
@@ -95,11 +138,116 @@ function EditUserProfile({ pfpurl, cancelFn }: EditUserProfileProps) {
         }
     }
 
+    function openUsernameChange() {
+        if (!authUser.verified) {
+            setUpdateMessage("You must verify your account first!")
+            return
+        }
+
+        setChangingUsername(true)
+    }
+
+    function handleEmailChange() {
+        if (checkEmail(newEmail)) {
+            changeEmailMutation.mutate(newEmail);
+        } else {
+            setUpdateMessage("Invalid email format, ex: yourname@domain.com")
+        }
+    }
+    function handleUsernameChange() {
+        if (checkUsername(newUsername)) {
+            changeUsernameMutation.mutate(newUsername)
+        } else {
+            setUpdateMessage("Invalid username, usernames must be between 2 and 64 characters")
+        }
+    }
+    function handlePasswordRequest() {
+        if (!authUser.verified) {
+            setUpdateMessage("You must verify your account first!")
+            return
+        }
+        setRequesedPassword(true)
+        changePasswordMutation.mutate(authUser.email)
+    }
+
     return (
+        <>
+        {
+            updateMessage != "" && <FadePopup key={updateMessage} text={updateMessage} duration={3.5}/>
+        }
         <div className="userProfileContainer">
             <h3>{authUser && authUser.username}<hr /></h3>
-            <div className="profilePictureContainer">
-                <ProfilePictureChanger imageURL={pfpurl} size={pfpSize} image={profileImage} setImage={setProfileImage} />
+            <div className="profileContainer">
+                <div className="profileTopHalf">
+                    <div className="profilePictureContainer">
+                        <ProfilePictureChanger imageURL={pfpurl} size={pfpSize} image={profileImage} setImage={setProfileImage} />
+                    </div>
+                    <div style={{flexDirection: "column", display: "flex", gap: "0", fontSize: "0.9rem"}}>
+                        <div>
+                            <span style={{fontWeight: "bold"}}>{authUser.username}</span>
+                            <button 
+                                className="floatingButton menuButton"
+                                onClick={openUsernameChange}
+                                style={{marginLeft: "0.5rem"}}
+                            >
+                                Change
+                            </button>
+                            {
+                                changingUsername && 
+                                <PopupMenu
+                                    title="Change Username"
+                                    text="Enter your new username (Warning! You can only change this once every 24 hours)"
+                                    onSubmit={handleUsernameChange}
+                                    onCancel={() => setChangingUsername(false)}
+                                >
+                                    <input 
+                                        type="text" 
+                                        className="smallTextInput" 
+                                        id="newUsername" 
+                                        value={newUsername} 
+                                        onChange={(e: ChangeEvent<HTMLInputElement>) => setNewUsername(e.target.value)}
+                                    />
+                                </PopupMenu>
+                            }
+                        </div>
+                        <button 
+                            className="floatingButton menuButton"
+                            onClick={handlePasswordRequest}    
+                        >
+                            Change Password
+                        </button>
+                        <div>
+                            {
+                                (authUser.email && authUser.email != "") && <span style={{fontWeight: "bold"}}>{authUser.email}</span>
+                            }
+                            <button 
+                                className="floatingButton menuButton"
+                                onClick={() => setChangingEmail(true)}
+                            >
+                                {
+                                    (authUser.email && authUser.email != "") ? "Change" : "Add Email"
+                                }
+                            </button>
+                            {
+                                changingEmail && 
+                                <PopupMenu
+                                    title={(authUser.email && authUser.email != "") ? "Change Email" : "Add Email"}
+                                    text="Enter your new email (You will have to verify your email address)"
+                                    onSubmit={handleEmailChange}
+                                    onCancel={() => setChangingEmail(false)}
+                                >
+                                    <input 
+                                        type="text" 
+                                        className="smallTextInput" 
+                                        id="newEmail" 
+                                        value={newEmail} 
+                                        onChange={(e: ChangeEvent<HTMLInputElement>) => setNewEmail(e.target.value)}
+                                    />
+                                </PopupMenu>
+                            }
+                        </div>
+                    </div>
+                </div>
             </div>
             <PreferenceMenu 
                 noteActivity={noteActivity}
@@ -121,6 +269,7 @@ function EditUserProfile({ pfpurl, cancelFn }: EditUserProfileProps) {
             Cancel
             </button>
         </div>
+        </>
     )
 }