+++ lifters-log-rewrite/src/types/UserData.ts
@@ -1,6 +1,8 @@
 type UserData = {
     username: string;
     userid: number;
+    verified: boolean;
+    email: string;
 }
 
 export default UserData;